requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
sqlalchemy==2.0.25
requests==2.31.0
apscheduler==3.10.4
